refactor: centralize data classes under types packages
