+++ src/anypod/cli/debug_downloader.py
@@ -13,8 +13,7 @@ from pathlib import Path
 from ..config import AppSettings
 from ..data_coordinator.downloader import Downloader
 from ..db import DownloadDatabase
-from ..db.download import Download
-from ..db.download_status import DownloadStatus
+from ..db.types import Download, DownloadStatus
 from ..exceptions import DatabaseOperationError, DownloadError
 from ..file_manager import FileManager
 from ..path_manager import PathManager
