+++ src/anypod/config/feed_config.py
@@ -12,38 +12,7 @@ from typing import Any
 from pydantic import BaseModel, Field, field_validator
 
 from ..ytdlp_wrapper.ytdlp_core import YtdlpCore
-from .podcast_categories import PodcastCategories
-from .podcast_explicit import PodcastExplicit
-
-
-class FeedMetadataOverrides(BaseModel):
-    """Podcast metadata overrides for RSS feed generation.
-
-    These values can be specified in the feed configuration to override
-    metadata that would otherwise be extracted from the yt-dlp output.
-    If not specified here, the system will attempt to extract these
-    values from the source content where possible (e.g., channel name
-    as title, channel description as description).
-
-    All fields are optional to allow selective overrides.
-    """
-
-    title: str | None = Field(None, description="Podcast title")
-    subtitle: str | None = Field(None, description="Podcast subtitle")
-    description: str | None = Field(None, description="Podcast description")
-    language: str | None = Field(
-        None, description="Podcast language (e.g., 'en', 'es')"
-    )
-    categories: PodcastCategories = Field(
-        default_factory=PodcastCategories,
-        description="Apple Podcasts category/categories (max 2)",
-    )
-    explicit: PodcastExplicit | None = Field(None, description="Explicit content flag")
-    image_url: str | None = Field(
-        None,
-        description="Podcast image URL, must be at least 1400x1400px, ideally 3000x3000px",
-    )
-    author: str | None = Field(None, description="Podcast author")
+from .types import FeedMetadataOverrides
 
 
 class FeedConfig(BaseModel):
