+++ src/anypod/config/types/podcast_explicit.py
@@ -1,10 +1,12 @@
+"""Podcast explicit content flags."""
+
 from enum import Enum
 from typing import Any
 
 from pydantic import GetCoreSchemaHandler
 from pydantic_core import CoreSchema, core_schema
 
-from ..db.sqlite_utils_core import register_adapter
+from ...db.sqlite_utils_core import register_adapter
 
 
 class PodcastExplicit(str, Enum):
