+++ src/anypod/data_coordinator/enqueuer.py
@@ -11,9 +11,8 @@ from typing import Any
 
 from ..config import FeedConfig
 from ..db import DownloadDatabase
-from ..db.download import Download
-from ..db.download_status import DownloadStatus
 from ..db.feed_db import FeedDatabase
+from ..db.types import Download, DownloadStatus
 from ..exceptions import (
     DatabaseOperationError,
     DownloadNotFoundError,
