+++ tests/anypod/db/test_download_db.py
@@ -10,8 +10,7 @@ from typing import Any
 import pytest
 
 from anypod.db import DownloadDatabase
-from anypod.db.download import Download
-from anypod.db.download_status import DownloadStatus
+from anypod.db.types import Download, DownloadStatus
 from anypod.exceptions import DatabaseOperationError, DownloadNotFoundError
 
 # --- Fixtures ---
