+++ tests/anypod/test_rss_feed.py
@@ -9,15 +9,14 @@ from xml.etree import ElementTree as ET
 
 import pytest
 
-from anypod.config import (
-    FeedConfig,
+from anypod.config import FeedConfig
+from anypod.config.types import (
     FeedMetadataOverrides,
     PodcastCategories,
     PodcastExplicit,
 )
 from anypod.db import DownloadDatabase
-from anypod.db.download import Download
-from anypod.db.download_status import DownloadStatus
+from anypod.db.types import Download, DownloadStatus
 from anypod.exceptions import DatabaseOperationError, RSSGenerationError
 from anypod.path_manager import PathManager
 from anypod.rss.rss_feed import RSSFeedGenerator
