+++ tests/anypod/ytdlp_wrapper/test_youtube_handler.py
@@ -8,9 +8,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from anypod.db.download import Download
-from anypod.db.download_status import DownloadStatus
-from anypod.db.source_type import SourceType
+from anypod.db.types import Download, DownloadStatus, SourceType
 from anypod.ytdlp_wrapper.base_handler import FetchPurpose
 from anypod.ytdlp_wrapper.youtube_handler import (
     ReferenceType,
